persist-txn: split TxnsCache into a new module
This commit is pure code movement. Everything in the new file is
copy+pasted, except I reordered the fn definitions a bit. All fixups,
including making things compile again are in the next commit.
